Move StandardIngestCoordinator.Builder into IngestCoordinator
